Bug 33260: (follow-up) Sync functionality and ordering of display
This patch syncs up the functionality and order of the notice displays.

We allow for resending of notices via the modal in the new tools page
and we add print in bulk to the members notices page.

We also fix the modal on the tools page to work with bootstrap 5
correctly.