Bug 33260: (follow-up) Limit printing to non-pending on members page
This patch prevent printing of notices in the 'pending' state.  This
ensures that email notices will not be skipped by process message queue
at it's next run as we set the 'sent' status on print with this
patchset.

It still allows for printing of sent and failed messages which I believe
is the purpose of the original patchset.