Bug 38712: order My Lists options alphabetically
In the top header navigation menu, sets the order of the 'My Lists' dropdown to be alphabetical.

Note: this patch was updated to also handle cases where $params->{search_attributes} is not defined,
and well as ordering public lists when the user is not logged in.

Test plan:

- on the Koha OPAC, navigate to /cgi-bin/koha/opac-shelves.pl
- create a few public and private lists
- open the 'My List' dropdown in the top navigation menu
- notice that the lists appear in the order they were created

- apply the patch

- open the 'My List' dropdown in the top navigation
- notice that the lists are now sorted alphabetically

fix: handle undefined search attributes

order list when logged out