Bug 38924: Overload old::checkout delete method for quotas
This overloads the delete method for old::checkouts to ensure that the
issue_id for quotas usage is nulled in the event that an old_checkout is
deleted.
Would like to make this the standard for all the tables we have had to
handle issue/old_issues for but is currently just hard-coded for quotas
as many of the other tables with issue_id columns are not nullable.