Ticket 24436: Update for more regular running of edi_cron
The edi cron on UAL runs every 10 minutes rather than once a day, thus
breaking some assumptions the original fix script had.

This patch updates the fixup script such that is uses edifact_messages
of the type 'QUOTE' with the statuses of 'received' or 'recmsg' and then
updates the status to 'post-$status' to track those quotes that have
been post processed.